+++ obot_g474/config/config_obot_g474_motor_pz2.cpp
@@ -264,14 +264,14 @@ void config_maintenance() {
         round_robin_logger.log_data(AMBIENT_TEMPERATURE_4_INDEX, Tambient4);
     }
     if(config::motor_encoder.crc_error_count_ > 100 || config::motor_encoder.error_count_ > 100 ||
-        config::motor_encoder.warn_count_ > pow(2,31)) {
+        config::motor_encoder.warn_count_ > powf(2,31)) {
             config::main_loop.status_.error.motor_encoder = true;
     }
     round_robin_logger.log_data(MOTOR_ENCODER_CRC_INDEX, config::motor_encoder.crc_error_count_);
     round_robin_logger.log_data(MOTOR_ENCODER_ERROR_INDEX, config::motor_encoder.error_count_);
 
     if(config::output_encoder_direct.crc_error_count_ > 100 || config::output_encoder_direct.error_count_ > 100 ||
-        config::output_encoder_direct.warn_count_ > pow(2,31)) {
+        config::output_encoder_direct.warn_count_ > powf(2,31)) {
             config::main_loop.status_.error.output_encoder = true;
     }
     round_robin_logger.log_data(OUTPUT_ENCODER_CRC_INDEX, config::output_encoder_direct.crc_error_count_);
